requirements.txt
contourpy==1.1.0
cycler==0.11.0
fonttools==4.40.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.25.0
packaging==23.1
Pillow==9.5.0
pyparsing==3.1.0
python-dateutil==2.8.2
pyusb==1.2.1
rti.connext @ file:///home/reidm/venv_test/Defender/app_files/rti.connext-7.1.0-cp310-cp310-linux_aarch64.whl
six==1.16.0
usb==0.0.83.dev0
